[rend2] Add special case for glow surfaces with sort < 2
In vanilla this leads to glow being drawn but not the actual stage. So we mask the first render target in this case to achieve the same look. Currently having flickering issues with surfaces nearly perpendicular to the viewer and enabled msaa. This sometimes leads to fireflys and I have no idea why. Is this a driver issue or is this intended behaviour? Need to look into that.