[renderers] Add ROM cvar to control MSAA framebuffer
When we changed r_ext_framebuffer_multisample to r_ext_multisample in rend2 we created an issue. The default framebuffer was multisampled as well as rend2 framebuffers. This is unnecessary overhead. Having a ROM cvar that is set by the renderer to tell the sdl window creation code to skip multisampling the default framebuffer is added. This is my prefered solution as we can use r_ext_multisample regardless of renderer. Renderers just need to tell the engine whether they do multisampling manually or if the default framebuffer should be multisampled.