[rend2] Add volumetric fog
This uses the lightgrid to create the illusion of volumetric fog. Requires a lightgrid per map obviously. When no lightgrid is found, it will use a default white 3d texture instead.
Since this can change the perception of fog tremendously there's a new worldspawn key "volumetricFogScale" to scale the volumetric fog density. Users can also set a default value if this key is not set via r_volumetricFogDefaultScale. r_volumetricFogScale can be used to determain a good value as it will additionaly scale the current scale value. The quality of the effect can be altered with the r_volumetricFogSamples cvar.